split version.go
