+++ lib/version_test.go
@@ -1,4 +1,4 @@
-package main
+package tss
 
 import (
 	"testing"
