+++ main.go
@@ -14,34 +14,10 @@ import (
 
 func init() {
 	flag.Usage = func() {
-		fmt.Fprintf(os.Stderr, `tss [-v] [-h]
+		fmt.Println(`tss [-v] [-h]
 
-Annotate stdin with timestamps per line.
-`)
+Annotate stdin with timestamps per line.`)
 	}
-}
-
-// tss build env
-var buildVersion string = "0.1.2-development"
-var buildCommit string = "aaeb6725631dcff02055855ee263ef5f45ed1eea-development"
-var buildDate string = "2020-12-28T11:01:32Z-development"
-
-// BuildEnvString returns tss build environment variables
-func BuildEnvString(goOS string, goArch string, goVersion string) string {
-	version := buildVersion
-	commit := buildCommit
-
-	date := buildDate
-	if date == "" {
-		date = "unknown"
-	}
-
-	return fmt.Sprintf(`TSS_BUILD_VERSION="%s"
-TSS_BUILD_COMMIT="%s"
-TSS_BUILD_DATE="%s"
-TSS_BUILD_GOOS="%s"
-TSS_BUILD_GOARCH="%s"
-TSS_BUILD_GOVERSION="%s"`, version, commit, date, goOS, goArch, goVersion)
 }
 
 func main() {
@@ -49,7 +25,7 @@ func main() {
 	v := flag.Bool("v", false, "Print the version string")
 	flag.Parse()
 	if *version || *v {
-		fmt.Println(BuildEnvString(runtime.GOOS, runtime.GOARCH, runtime.Version()))
+		fmt.Println(tss.BuildEnvString(runtime.GOOS, runtime.GOARCH, runtime.Version()))
 		os.Exit(0)
 	}
 	if _, err := tss.Copy(os.Stdout, os.Stdin); err != nil {
